Split c file into more manageable files
